+++ starters/pom.xml
@@ -15,7 +15,7 @@
     </parent>
 
     <modules>
-        <module>spring-web-starters</module>
+        <module>spring-starters</module>
     </modules>
 
 </project>
