+++ starters/spring-starters/pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>spring-web-starters</artifactId>
+    <artifactId>spring-starters</artifactId>
     <name>spring-web-starters</name>
     <packaging>pom</packaging>
 
@@ -14,6 +14,7 @@
     </parent>
 
     <modules>
+        <module>spring-examples</module>
         <module>spring-web-c7</module>
     </modules>
