+++ starters/spring-starters/spring-web-c7/pom.xml
@@ -9,7 +9,7 @@
 
     <parent>
         <groupId>io.miragon.miranum</groupId>
-        <artifactId>spring-web-starters</artifactId>
+        <artifactId>spring-starters</artifactId>
         <version>0.1.0-SNAPSHOT</version>
     </parent>
 
@@ -20,6 +20,12 @@
         </dependency>
 
         <!--Miranum-->
+        <dependency>
+            <groupId>io.miragon.miranum</groupId>
+            <artifactId>spring-examples</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
         <dependency>
             <groupId>io.miragon.miranum</groupId>
             <artifactId>message-adapter-cloudevents</artifactId>
