+++ starters/spring-starters/spring-web-c7/src/main/java/io/miragon/miranum/starter/spring/web/c7/SpringWebC7Application.java
@@ -1,0 +1,12 @@
+package io.miragon.miranum.starter.spring.web.c7;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringWebC7Application {
+
+    public static void main(final String[] args) {
+        SpringApplication.run(SpringWebC7Application.class, args);
+    }
+}
