+++ starters/spring-starters/spring-web-c7/src/main/resources/application.properties
